refactor(mongodb): split bulkWrite span into its own child plugin
The bulkWrite span lived inside the query plugin's constructor, which
hand-wired its start/finish/error channels and re-implemented the
parent-store restore that `OutboundPlugin.bindFinish` already provides.
Move it into a `MongodbCoreBulkWritePlugin` (operation `bulkwrite`, so it
resolves the `apm:mongodb:bulkwrite:*` channels) that inherits `finish`,
`error`, and the finish-time parent restore from the tracing/outbound base
classes, and wire both spans through a `CompositePlugin`. It extends the
query plugin to reuse the shared id, component, and collection-stripping
`getPeerService`.

Adding a third MongoDB logical span is now another small child plugin
rather than more ad hoc constructor wiring, and the lifecycle contract is
enforced by the base classes instead of by convention.
